+++ stat_other.go
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 // this excludes all known build targets, so it will only build on newly introduced systems:
-//go:build !(aix || darwin || dragonfly || freebsd || illumos || linux || netbsd || openbsd || plan9 || solaris || wasm || windows)
+//go:build !(aix || darwin || dragonfly || freebsd || js || illumos || linux || netbsd || openbsd || plan9 || solaris || wasip1 || windows)
 
 package compat