+++ stat_tinygo_wasip1.go
@@ -1,0 +1,16 @@
+// SPDX-FileCopyrightText: Copyright © 2025 Ross Smith II <ross@smithii.com>
+// SPDX-License-Identifier: MIT
+
+//go:build wasip1 && tinygo
+
+package compat
+
+import (
+	"time"
+)
+
+// Not supported: ATime | BTime | CTime.
+const supported SupportedType = Links | UID | GID
+
+func (fs *fileStat) times() {
+}