+++ umask_unix.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright © 2025 Ross Smith II <ross@smithii.com>
 // SPDX-License-Identifier: MIT
 
-//go:build aix || darwin || dragonfly || freebsd || illumos || linux || netbsd || openbsd || solaris || wasm
+//go:build aix || darwin || dragonfly || freebsd || illumos || js || linux || netbsd || openbsd || solaris || (wasip1 && !tinygo)
 
 package compat
 
@@ -23,7 +23,7 @@ func init() {
 // On Windows, the initial umask value is 022 octal, and can be changed by
 // setting environmental variable UMASK, to an octal value. For example:
 // `set UMASK=022` . Leading zeros and 'o's are allowed, and ignored.
-// On Plan9, the function does nothing, and always returns zero.
+// On Plan9 and Wasip1, the function does nothing, and always returns zero.
 func Umask(newMask int) int {
 	umaskMutex.Lock()
 	defer umaskMutex.Unlock()
@@ -33,6 +33,7 @@ func Umask(newMask int) int {
 }
 
 // GetUmask returns the current umask value.
+// On Plan9 and Wasip1, the function always returns zero.
 func GetUmask() int {
 	umaskMutex.Lock()
 	defer umaskMutex.Unlock()